% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{analyze,pathway-method}
\alias{analyze,pathway-method}
\title{analyze pathway network properties}
\usage{
\S4method{analyze}{pathway}(object)
}
\arguments{
\item{\code{pathway}}{object}
}
\value{
\code{data.frame} consisting of
   \item{id} pathway id,
   \item{vcount} number of genes,
   \item{ecount} number of links,
   \item{inh_ecount} number of inhibition links,
   \item{density} network density,
   \item{av_deg} average degree,
   \item{inh_deg} average degree of inhibition links,
   \item{diam} network diamter,
   \item{trans} transitivity, and
   \item{s_trans} signed transitivity (Kunegis et al., 2009).
}
\description{
analyze pathway network properties
}
\examples{
data(hsa04710)
summary(hsa04710)
analyze(hsa04710)
}
\author{
Juliane Manitz
}
\references{
Details to the computation and interpretation can be found in:
  Kolaczyk, E. D. (2009). Statistical analysis of network data: methods and models. Springer series in statistics. Springer.
  Kunegis, J., A. Lommatzsch, and C. Bauckhage (2009). The slashdot zoo: Mining a social network with negative egdes. In Proceedings of the 18th international conference on World wide web, pp. 741–750. ACM Press.
}

