% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{anno}
\alias{anno}
\title{Example annotation file for three pathways.}
\format{A \code{data frame} with 4056 rows and 5 variables:
\describe{
 \item{pathway}{includes KEGG identifiers of three example pathways}
 \item{gene}{names of genes in the pathways}
 \item{chr}{specifies the chromosom}
 \item{snp}{includes rs-numbers of example SNPs}
 \item{position}{gives positions of example SNPs}
}}
\source{
simulated data
}
\usage{
data(anno)
}
\description{
A dataset containing an annotation example for 4056 SNPs in three different
pathways.
}
\keyword{datasets}

