% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lkmt.r
\docType{methods}
\name{davies_test,matrix-method}
\alias{davies_test,matrix-method}
\title{Calculates the p-value for a kernel matrix using davies method}
\usage{
\S4method{davies_test}{matrix}(x1, x2)
}
\arguments{
\item{x1}{A \code{\link{matrix}} which is the
similarity matrix calculated for the pathway to be tested.}

\item{x2}{A \code{glm} object of the nullmodel with fixed effects
covariates included, but no genetic random effects.}
}
\value{
A \code{list} including the test results of the pathway.
}
\description{
This function evaluates a pathways influence on an individuals probability
of beeing a case using the logistic kernel machine test. P-values are
determined using the method described by Davies as implemented in the
 function \code{davies()} from package \code{CompQuadForm}.
}
\author{
Stefanie Friedrichs
For details on the p-value calculation see
\itemize{
\item Davies R: Algorithm as 155: the distribution of a linear combination of
     chi-2 random variables. J R Stat Soc Ser C 1980, 29:323-333.
}
}

