\name{definition}
\alias{definition}
\title{
Number of genes of SNPs. 
}
\description{
Calculates with help of annotation file number of genes and SNPs. 
}
\usage{
definition(object)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{object}{
  Object of class \code{GWASdata}. 
}
}
\value{
\item{tab}{
  Returns vector which contains number of genes and SNPs.
  }
}
\author{
Juliane Manitz, Stefanie Friedrichs, Patricia Burger, Benjamin Hofner
}

\seealso{
\code{\link{GWASdata}} for more information about \code{object}. 
}

\keyword{definition}

