\name{EffectiveNumberSNPs}
\alias{EffectiveNumberSNPs}
\title{
Effective number of SNPs
}
\description{
This function calculates the effective number of SNPs. 
}
\usage{
EffectiveNumberSNPs(object, genes.ids)
}
\arguments{
  \item{object}{
  Object of the type \code{GWASdata}. 
}
  \item{genes.ids}{
  Gene IDs 
}
}
\value{
\item{effectivesnps}{Effective number of SNPs.}
}

\author{
Juliane Manitz, Stefanie Friedrichs, Patricia Burger, Benjamin Hofner
}

\seealso{
\code{\link{GWASdata}} for more information about \code{object}. 
}
\keyword{EffectiveNumberSNPs}
