% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\name{ff-class}
\alias{ff}
\alias{ff-class}
\title{"ff" class for memory-efficient storage of genotype matrix on disk and fast access}
\description{
"ff" class for memory-efficient storage of genotype matrix on disk and fast access
}
\seealso{
ff::ffdf ff::read.table.ff
}

