% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\name{ffdf-class}
\alias{ffdf}
\alias{ffdf-class}
\title{"ffdf" class for memory-efficient storage of genotype data frame on disk and fast access}
\description{
"ffdf" class for memory-efficient storage of genotype data frame on disk and fast access
}
\seealso{
ff::ffdf ff::read.table.ff
}

