% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\name{gene_name_number}
\alias{gene_name_number}
\title{Function to get genes names and numbers from kegg}
\usage{
gene_name_number(id)
}
\arguments{
\item{id}{A \code{character} hsa identifier of the pathway for which gene
infomation should be extracted as used in KEGG database ('hsa00100').}
}
\value{
A \code{data.frame} listing the genes included in the pathway with
their names as well as numbers used in KEGG database.
}
\description{
This function extracts for a particular pathway all included genes and the
numbers they are given within the corresponding KGML pathway file.
}
\examples{
#pathway_info("hsa04710") #### DOES NOT WORK ###
}
\author{
Stefanie Friedrichs
}

