% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{GeneSNPsize,GWASdata-method}
\alias{GeneSNPsize,GWASdata-method}
\title{Counts the number of Genes and SPNs in each pathway}
\usage{
\S4method{GeneSNPsize}{GWASdata}(object)
}
\arguments{
\item{gwas}{\code{GWASdata} object.}
}
\value{
Creates a list of pathway names with numbers of snps and genes in the pathway.
}
\description{
Counts the number of Genes and SPNs in each pathway
}
\examples{
#data(gwas)### ERROR
#GeneSNPsize(gwas) ### ERROR
}
\author{
Juliane Manitz
}

