\name{GeneSNPsize-methods}
\docType{methods}
\alias{GeneSNPsize-methods}
\alias{GeneSNPsize,GWASdata-method}
\title{Methods for Function \code{GeneSNPsize}}
\description{
Methods for function \code{GeneSNPsize}
}
\section{Methods}{
\describe{
Calculates the GeneSNPsize. 
\item{\code{signature(object = "GWASdata")}}{
%%  ~~describe this method here~~
}
}}
\keyword{methods}
