% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{geno}
\alias{geno}
\title{Example genotypes for 50 individuals.}
\format{
A \code{matrix} with 5 rows and 4056 columns:
\describe{
 each entry in the matrix represents a simulated minor allele count for the
corresponding SNP and individual.
}
}
\source{
simulated data
}
\usage{
data(geno)
}
\description{
A matrix containing example genotypes for 4056 SNPs of 50 individuals. 
Column names give the rs-numbers of 4056 example SNPs, row names the 
identifiers of 50 example individuals.
}
\examples{
data(geno)
head(geno)
# create gwas object
data(pheno)
data(anno)
gwas <- new('GWASdata', pheno=pheno, geno=geno, anno=anno, desc="some study") 
}
\keyword{datasets}
