% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.r
\docType{methods}
\name{get_ana,data.frame-method}
\alias{get_ana,data.frame-method}
\title{Produce middle part of network kernel (for internal use)}
\usage{
\S4method{get_ana}{data.frame}(x, SNPset, pathway)
}
\arguments{
\item{x}{\code{data.frame} with annotation information as returned from
\code{get_anno}}

\item{SNPset}{vector with SNPs to be analyzed}

\item{pathway}{pathway object}
}
\value{
matrix ANA' for inner part of network kernel
}
\description{
Produce middle part of network kernel (for internal use)
}
\author{
Juliane Manitz, Saskia Freytag, Stefanie Friedrichs
}
\keyword{internal}
