% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\name{get_ana}
\alias{get_ana}
\title{Produce middle part of Network Kernel (for internal use)}
\usage{
get_ana(anno, SNPset, pathway)
}
\arguments{
\item{anno}{\code{data.frame} with annotation information as returned from
\cite{\link{get_anno}}}

\item{SNPset}{vector with SNPs to be analyzed}

\item{pathway}{pathway object}
}
\value{
matrix ANA' for inner part of network kernel
}
\description{
Produce middle part of Network Kernel (for internal use)
}
\author{
Juliane Manitz, Saskia Freytag, Stefanie Friedrichs
}

