% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{get_anno,snp_info,pathway_info-method}
\alias{get_anno,snp_info,pathway_info-method}
\title{A function to create the annotation for a GWASdata object, in which
SNPs are assigned to pathways via gene membership.}
\usage{
\S4method{get_anno}{snp_info,pathway_info}(object1, object2, ...)
}
\arguments{
\item{object1}{A \code{snp_info} object with SNP information as returned by
the \code{\link{snp_info}} function. The included \code{data frame} contains
the columns "chr", "position" and "rsnumber".}

\item{object2}{A \code{pathway_info} object with information on genes
contained in pathways. It is created by the \code{\link{pathway_info}}
function and contains a \code{data frame} with columns
the columns "pathway", "gene_start", gene_end", "chr", "gene".}

\item{...}{further arguments can be added.}
}
\value{
A \code{data.frame} including mapping SNPs to genes and genes to
pathways. It includes the columns "pathway", "gene", "chr", "snp" and
"position".
}
\description{
This function combines a \code{snp_info} and a \code{pathway_info} object
into an annotation \code{data.frame} used for pathway analysis on GWAS.
}
\examples{
data(hsa04022_info)
data(rs10243170_info)
}
\author{
Stefanie Friedrichs, Saskia Freytag, Ngoc-Thuy Ha
}
\seealso{
\code{\link{snp_info}}, \code{\link{pathway_info}}
}

