% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\name{get_anno}
\alias{get_anno}
\title{Create annotation for GWASdata object}
\usage{
get_anno(snp_info, pathway_info, ...)
}
\arguments{
\item{snp_info}{A \code{data frame} with SNP information as returned by
\code{\link{snp_info}}. The \code{data frame} has to contain columns "chr",
"position" and "rsnumber".}

\item{pathway_info}{A \code{data frame} with information on the genes forming
the pathway. Output from \code{\link{pathway_info}}.}

\item{...}{further arguments can be added.}
}
\value{
A \code{data.frame} mapping SNPs to genes and genes to pathways. Includes
the columns "pathway", "gene", "chr", "snp" and "position".
}
\description{
This function gives for a vector of SNPs the position of each SNP as
extracted from the Ensemble database. The database is accessed via the
R-package \code{biomaRt}.
}
\examples{
#### missing example #### <FIXME>
}
\author{
Stefanie Friedrichs, Saskia Freytag, Ngoc-Thuy Ha
}

