% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{get_genes,pathway-method}
\alias{get_genes,pathway-method}
\title{helper function extracting the gene names in a pathway}
\usage{
\S4method{get_genes}{pathway}(object)
}
\arguments{
\item{pathway}{pathway object}
}
\value{
a vector of character containing gene names
}
\description{
helper function extracting the gene names in a pathway
}
\examples{
data(hsa04710)
get_genes(hsa04710)
}
\author{
Juliane Manitz, benjamin Hofner
}

