% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{get_network_matrix,character-method}
\alias{get_network_matrix,character-method}
\title{Function to calculate the network matrix}
\usage{
\S4method{get_network_matrix}{character}(x, directed = TRUE,
  keep.kgml = TRUE, method = "KEGG")
}
\arguments{
\item{x}{A \code{character} identifying the pathway for which gene infomation
should be extracted. Here KEGG IDs ('hsa00100') are used.}

\item{directed}{A \code{logic} argument, stating whether the networkmatrix
should be returned directed (\code{TRUE}) or undirected (\code{FALSE}).}

\item{keep.kgml}{A \code{logic} argument, specifying whether the downloaded
KEGG kgml file of the pathway should be kept in the working directory after
calculation of the network matrix. For (\code{FALSE}) the file is deleted,
for (\code{TRUE}) not.}

\item{method}{A \code{character} stating from where the pathway data should
be downloaded. Currently only 'KEGG' for the Kegg database is available.}
}
\value{
A \code{matrix} representing the interaction network in the pathway.
}
\description{
This function creates the networkmatrix representing the gene-gene interaction
structure within a particular pathway. In this process a KEGG kgml file is
downloaded and saved in the working directory.
}
\author{
Stefanie Friedrichs, Patricia Burger
}

