% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{get_network_matrix,pathway-method}
\alias{get_network_matrix}
\alias{get_network_matrix,pathway-method}
\title{Function to calculate the network \code{matrix} for a \code{\link{pathway}} object}
\usage{
\S4method{get_network_matrix}{pathway}(object, directed = TRUE)
}
\arguments{
\item{directed}{A \code{logical} argument, stating whether the network matrix 
should return directed (\code{TRUE}) or undirected (\code{FALSE}) links.}
}
\value{
\code{get_network_matrix} returns the modified \code{\link{pathway}} object, where the slots \code{'adj'} and \code{'sign'} are altered according to the downloaded information in the KEGG kgml file.
}
\description{
\code{get_network_matrix} creates the adjacency matrix representing the gene-gene interaction structure within a particular \code{\link{pathway}}. Note that a 
KEGG kgml file is downloaded and saved in the working directory.
}
\author{
Juliane Manitz, Stefanie Friedrichs, Patricia Burger
}

