% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\name{get_network_matrix}
\alias{get_network_matrix}
\title{Function to calculate the network matrix}
\usage{
get_network_matrix(id, directed, ...)
}
\arguments{
\item{id}{A \code{character} identifying the pathway for which gene infomation
should be extracted. Here KEGG IDs ('hsa00100') are used.}

\item{directed}{A \code{logic} argument, stating whether the networkmatrix
should be returned directed (\code{TRUE}) or undirected (\code{FALSE}).}

\item{...}{additional aguments can be specified.}
}
\value{
A \code{matrix} representing the interaction network in the pathway.
}
\description{
This function creates the networkmatrix representing the genge-gene interaction
structure within a particular pathway.
}
\examples{
# get_network_matrix("hsa04710", TRUE) ### DOES NOT WORK
}
\author{
Stefanie Friedrichs
}

