% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas}
\alias{gwas}
\title{Example GWASdata object.}
\format{An object of class \code{GWASdata}:
\describe{
 \item{geno}{contains example genotypes}
 \item{anno}{example annotation for threee pathways}
 \item{pheno}{exemplary genotypes for the iondividuals listed in geno}
 \item{desc}{a description of the GWAS study, here 'example study'}
}}
\source{
simulated data
}
\usage{
data(gwas)
}
\description{
An object of type GWASdata containing the example files for
annotation, phenotypes and genotypes.
}
\keyword{datasets}

