% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{GWASdata,ANY-method}
\alias{GWASdata,ANY-method}
\alias{GeneSNPsize,GWASdata,ANY-method}
\alias{GeneSNPsize,GWASdata-method}
\alias{show,GWASdata,ANY-method}
\alias{show,GWASdata-method}
\alias{summary,GWASdata,ANY-method}
\alias{summary,GWASdata-method}
\title{\code{'GWASdata'} is a GWASdata object constructor}
\usage{
\S4method{GWASdata}{ANY}(geno, pheno = NULL, anno = NULL, desc = NULL)

\S4method{show}{GWASdata}(object)

\S4method{summary}{GWASdata}(object)

\S4method{GeneSNPsize}{GWASdata}(object)
}
\arguments{
\item{geno}{an \code{ff} data frame including genotype information.}

\item{pheno}{A \code{data.frame} specifying individual IDs, phenotypes and
covariates to be included in the regression model e.g. ID, pheno, sex,
pack.years. Note: IDs have to be in the first column!}

\item{anno}{a \code{data.frame} mapping SNPs to genes and genes to
pathways. Needs to include the columns 'pathway' (pathway ID, e.g. hsa
number from KEGG database), 'gene' (gene name (hgnc_symbol)), 'chr'
(chromosome), 'snp' (rsnumber) and 'position' (base pair position of SNP).}

\item{desc}{A \code{character} giving the GWAS description, e.g. name of study

#}

\item{object}{A \code{GWASdata} object.}

\item{object}{A \code{GWASdata} object.}

\item{object}{A \code{GWASdata} object.}
}
\value{
\code{show} This function shows the phenotype information and the object description included in a \code{\link{GWASdata}} object.

\code{summary} This function gives an overview about the information included in a \code{\link{GWASdata}} object. Summary statistics for phenotype and genotype data are calculated.

\code{GeneSNPsize} Creates a list of pathway names with numbers of snps and genes in the pathway.
}
\description{
\code{'GWASdata'} is a GWASdata object constructor

\code{show} Shows basic information on \code{GWASdata} object

\code{summary} Summarizes the content of a \code{GWASdata} object

\code{GeneSNPsize} Counts the number of Genes and SPNs in each pathway
}
\examples{
data(gwas)
show(gwas)

#
# data(gwas)   #### ERROR
# summary(gwas) ### ERROR

#
#data(gwas)### ERROR
#GeneSNPsize(gwas) ### ERROR

#
}
\author{
Juliane Manitz

Juliane Manitz

Juliane Manitz

Juliane Manitz
}

