\name{GWASdata}
\alias{GWASdata}
\title{
New GWASdata object
}
\description{
Creates new object of class \code{GWASdata} with slots \code{pheno}, \code{geno} and \code{desc}. 
}
\usage{
GWASdata(...)
}
\arguments{
  \item{\dots}{
  Pheno, geno and desc. 
}
}

\author{
Juliane Manitz, Stefanie Friedrichs, Patricia Burger, Benjamin Hofner
}

\seealso{
For class information see \code{\link{GWASdata-class}}.
}
\keyword{GWASdata}

