% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsa04022_info}
\alias{hsa04022_info}
\title{Example \code{\link{pathway_info}} object for \code{\link{pathway}} hsa04022.}
\format{A \code{\link{pathway_info}} object including information on 163 genes.
\describe{
 \item{info}{a \code{data frame} including information on genes included in 
pathway. Has columns 'pathway', 'gene_start', 'gene_end', 'chr', and 'gene'}
}}
\source{
Ensembl extract
}
\usage{
data(hsa04022_info)
}
\description{
An object of class \code{\link{pathway_info}} for the \code{\link{pathway}}
with KEGG identifier hsa04020.
}
\examples{
\dontrun{
pathway_info('hsa04020')
}
}
\keyword{datasets}
