\name{kangar00-package}
\alias{kangar00-package}
\alias{kangar00}
\docType{package}
\title{
kangar00 package
}
\description{
Kernle methods for GWAS with pathways
}
\details{
\tabular{ll}{
Package: \tab kangar00\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-02-04\cr
License: \tab What license is it under?\cr
Depends: \tab methods\cr
}

}
\author{
Who wrote it

Maintainer: Who to complain to <yourfault@somewhere.net>

}
\references{

}

\keyword{ package }
\seealso{

}
\examples{

}
