\name{kernel.adj}
\alias{kernel.adj}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
Size-adjusted kernel
}
\description{
Calculates the size-adjusted kernel.
}
\usage{
kernel.adj(GWASdata, pathway)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{GWASdata}{
  Object of the class \code{GWASdata}. 
}
  \item{pathway}{
  Object of the class \code{pathway}
}
}

\value{
Returns object of class kernel. 
}

\seealso{
\code{\link{GWASdata}}, \code{\link{pathway}}, \code{\link{kernel}}
}


\keyword{kernel.adj}
\keyword{SizeAdjusedKernel}% __ONLY ONE__ keyword per line
