\name{kernel.net}
\alias{kernel.net}
\title{
Network-based kernel
}
\description{
Calculates a network-based kernel
}
\usage{
kernel.net(data, pathway)
}

\arguments{
  \item{data}{
  Object of the class \code{GWASdata}.
}
  \item{pathway}{
  Object of the class \code{pathway}.
}
}

\value{
Returns object of class kernel. 
}

\author{
Juliane Manitz, Stefanie Friedrichs, Patricia Burger, Benjamin Hofner
}

\keyword{NetworkKernel}

