% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lkmt.r
\name{lkmt}
\alias{lkmt}
\title{An S4 class defining an object to represent the variance component test.}
\usage{
lkmt(formula, kernel, GWASdata, ...)
}
\description{
the variance component test.
matrix of the individuals based on which the pathways influence is evaluated.
which the test is conducted.
ttest statistic.
in the variance component test.
}
\author{
Juliane Manitz, Stefanie Friedrichs
}

