% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\name{make_posdev}
\alias{make_posdev}
\title{Adjust matrix to be positive definite}
\usage{
make_posdev(N)
}
\arguments{
\item{N}{A kernelmatrix.}
}
\value{
The matrix N, if it is positive definite and the closes positive
definite matrix if N is not positive definite.
For more details check the references.
}
\description{
Adjust matrix to be positive definite
}
\author{
Juliane Manitz, Saskia Freytag, Stefanie Freidrichs
}
\references{
\itemize{
 \item Freytag S, Manitz J, Schlather M, Kneib T, Amos CI, Risch A, Chang-Claude J, Heinrich J, Bickeboeller H: A network-based kernel machine test for the identification of risk pathways in genome-wide association studies. Hum Hered. 2013, 76(2):64-75.
}
}

