% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\docType{methods}
\name{make_psd,matrix-method}
\alias{make_psd,matrix-method}
\title{Adjust network matrix to be positive semi-definite}
\usage{
\S4method{make_psd}{matrix}(x, eps = sqrt(.Machine$double.eps))
}
\arguments{
\item{x}{matrix specifying the network adjacency matrix.}

\item{eps}{numeric, tolance for smallest eigenvalue adjustment}
}
\value{
The matrix x, if it is positive definite and the closest positive semi-definite matrix if x is not positive semi-definite.
}
\description{
Adjust network matrix to be positive semi-definite
}
\details{
<FIXME> Add formula. For more details check the references.
}
\author{
Juliane Manitz, Saskia Freytag, Stefanie Friedrichs
}
\references{
\itemize{
 \item Freytag S, Manitz J, Schlather M, Kneib T, Amos CI, Risch A, Chang-Claude J, Heinrich J, Bickeboeller H: A network-based kernel machine test for the identification of risk pathways in genome-wide association studies. Hum Hered. 2013, 76(2):64-75.
}
}

