% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{class}
\name{pathway-class}
\alias{pathway}
\alias{pathway-class}
\title{An S4 class to represent a gene-gene interaction network}
\description{
An S4 class to represent a gene-gene interaction network
}
\section{Slots}{

\describe{
\item{\code{id}}{a character repesenting the pathway id, e.g. hsa00100}

\item{\code{adj}}{a matrix respresenting the network adjacency matrix (1 interaction, 0 otherwise)}

\item{\code{sign}}{a numeric vector indicating the interaction type for each link (1 activation, -1 inhibition)}
}}
\author{
Juliane Manitz, Stefanie Friedrichs
}
\seealso{
\code{\link{}}
}

