% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{pathway-method}
\alias{get_genes}
\alias{get_genes,pathway-method}
\alias{pathway-method}
\alias{pathway2igraph}
\alias{pathway2igraph,pathway-method}
\alias{sample_genes}
\alias{sample_genes,pathway-method}
\alias{show}
\alias{summary}
\alias{summary,pathway-method}
\title{Methods for objects from S4 class 'pathway'}
\usage{
\S4method{show}{pathway}(object)

\S4method{summary}{pathway}(object)

\S4method{pathway2igraph}{pathway}(object)

\S4method{get_genes}{pathway}(object)

\S4method{sample_genes}{pathway}(object, no = 3)
}
\arguments{
\item{object}{pathway object}

\item{no}{a numeric constant specifying the number of genes to be sampled, default is 3}
}
\description{
Methods for objects from S4 class 'pathway'

\code{show} displays the pathway object briefly

\code{summary} generates a pathway object summary including basic network properties.

\code{pathway2igraph} converts a \code{\link{pathway}} object into an \code{\link{igraph}} object with edge attribute \code{sign}

\code{get_genes} is a helper function that extracts the gene names in a pathway and returns a vector of character containing gene names

\code{sample_genes} function randomly selects effect genes in pathway and returns a vector of length \code{no} with vertex id's of sampled genes
}
\examples{
data(hsa04710)
hsa04710
summary(hsa04710)
# convert to igraph object
str(hsa04710)
g <- pathway2igraph(hsa04710)
str(g)
# extract gene names from pathway
get_genes(hsa04710)
# sample effect genes
sample3 <- sample_genes(hsa04710, no = 3)
plot(hsa04710, highlight.genes = sample3)
sample10 <- sample_genes(hsa04710, no = 10)
plot(hsa04710, highlight.genes = sample10)
}
\author{
Juliane Manitz
}
\seealso{
\code{\link{pathway-class}}, \code{\link{analyze}}, \code{\link{plot-pathway}}
}

