% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\name{pathway2igraph}
\alias{pathway2igraph}
\title{converts a \code{\link{pathway}} object to an \code{\link{igraph}} object}
\usage{
pathway2igraph(pathway)
}
\arguments{
\item{pathway}{pathway object}
}
\value{
This function returns an \code{\link{igraph}} object with edge attribute \code{sign}
}
\description{
converts a \code{\link{pathway}} object to an \code{\link{igraph}} object
}
\examples{
data(hsa04710)
str(hsa04710)
g <- pathway2igraph(hsa04710)
str(g)
}
\author{
Juliane Manitz
}

