% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{get_genes,pathway-method}
\alias{get_genes,pathway,ANY-method}
\alias{get_genes,pathway-method}
\title{\code{get_genes} is a helper function that extracts the gene names in a pathway and returns a vector of character containing gene names}
\usage{
\S4method{get_genes}{pathway}(object)
}
\description{
\code{get_genes} is a helper function that extracts the gene names in a pathway and returns a vector of character containing gene names
}
\examples{
# extract gene names from pathway
get_genes(hsa04710)
}

