% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{pathway_info,character-method}
\alias{pathway_info,character-method}
\title{Get information on genes in a pathway}
\usage{
\S4method{pathway_info}{character}(x)
}
\arguments{
\item{x}{A \code{character} identifying the pathway for which gene infomation
should be extracted. Here KEGG IDs ('hsa00100') are used.}
}
\value{
A \code{data.frame} including as many rows as genes appear in the
pathway. for each gene its name, the start and end point and the chromosome
it lies on are given.
}
\description{
This function lists all genes formig a particular pathway. Start and end
positions of these genes are extracted from the Ensemble database. The
database is accessed via the R-package \code{biomaRt}.
}
\examples{
pathway_info("hsa04022")
}
\author{
Stefanie Friedrichs
}

