% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{show,pathway_info-method}
\alias{show,pathway_info-method}
\alias{summary,pathway_info-method}
\title{\code{show} Shows basic information on \code{pathway_info} object}
\usage{
\S4method{show}{pathway_info}(object)

\S4method{summary}{pathway_info}(object)
}
\arguments{
\item{object}{An object of class \code{\link{pathway_info}}.}

\item{object}{An object of class \code{\link{pathway_info}}.}
}
\value{
\code{show} Basic information on \code{pathway_info} object.

\code{summary} Summarized information on \code{pathway_info} object.
}
\description{
\code{show} Shows basic information on \code{pathway_info} object

\code{summary} Summarizes information on \code{pathway_info} object
}
\examples{
data(hsa04022_info)
hsa04022_info
data(hsa04022_info)
summary(hsa04022_info)
}
\author{
Stefanie Friedrichs

Stefanie Friedrichs
}

