% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pheno}
\alias{pheno}
\title{Example phenotype file for 50 individuals.}
\format{A \code{data frame} with 50 rows and 3 variables:
\describe{
 \item{pheno}{includes the case-control status for each individual, coded as
   1(case) or 0 (control)}
 \item{sex}{includes gender information for the 50 individuals, coded as 1 
   (male) or 0 (female)}
 \item{age}{numerical value giving the persons age}
}}
\source{
simulated data
}
\usage{
data(pheno)
}
\description{
A dataset containing simulated example phenotypes for 50 individuals 
row names include the identifiers of 50 example individuals.
}
\examples{
data(pheno)
head(pheno)
# create gwas object
data(geno)
data(anno)
gwas <- new('GWASdata', pheno=pheno, geno=geno, anno=anno, desc="some study") 
}
\keyword{datasets}
