% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{plot,pathway,missing-method}
\alias{plot,pathway,missing-method}
\title{plot pathway object}
\usage{
\S4method{plot}{pathway,missing}(x, y = NA, highlight.genes = NULL,
  gene.names = c("legend", "nodes", NA), main = NULL, asp = 0.95,
  vertex.size = 11, vertex.color = "khaki1", vertex.label.cex = 0.8,
  edge.width = 2, edge.color = "olivedrab4", ...)
}
\arguments{
\item{x}{pathway object}

\item{y}{missing (placeholder)}

\item{highlight.genes}{vector of gene names or node id's, which should be highlighted in a different color, default is \code{NULL} so that no genes are highlighted}

\item{gene.names}{character indicating whether the genes names should appear in a legend (\code{'legend'}), as vertex label (\code{'nodes'}), or should be omitted (\code{NA})}

\item{main}{optional overall main title, default is \code{NULL}, which uses the pathway id}

\item{asp}{a numeric constant, which gives the aspect ratio parameter for plot, default is 0.95}

\item{vertex.size}{a numeric constant specifying the vertex size, default is 11}

\item{vertex.color}{a character or numeric constant specifying the vertex color, default is 'khaki1'}

\item{vertex.label.cex}{a numeric constant specifying the the vertex label size, default is 0.8,}

\item{edge.width}{a numeric constant specifying the edge width, default is 2}

\item{edge.color}{a character or numeric constant specifying the edge color, default is 'olivedrab4'}

\item{...}{further arguments specifying plotting options in \code{\link{plot.igraph}}}
}
\description{
plot pathway object
}
\examples{
data(hsa04710)
summary(hsa04710)
plot(hsa04710)
sample3 <- sample_genes(hsa04710, no = 3)
plot(hsa04710, highlight.genes = sample3)
}
\author{
Juliane Manitz, Saskia Freytag
}

