% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{read_geno,ANY-method}
\alias{read_geno,ANY-method}
\title{read genotype data from file to ff_matrix object, which can be passed to a  GWASdata object}
\usage{
\S4method{read_geno}{ANY}(path)
}
\arguments{
\item{path}{character, which contains the path to the txt file

import data.table ff
export}
}
\description{
read genotype data from file to ff_matrix object, which can be passed to a  GWASdata object
}

