% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{read_geno,ANY-method}
\alias{read_geno,ANY-method}
\title{read genotype data from file to bigmemory object, which can be passed to a  GWASdata object}
\usage{
\S4method{read_geno}{ANY}(file.path, save.path = NULL, sep = " ",
  header = TRUE, ...)
}
\arguments{
\item{file.path}{character, which contains the path to the data file to be read}

\item{save.path}{character, which contains the path for the backingfile}

\item{sep}{character. A field delimeter. See \code{bigmemory::read.big.matrix} for
details.}

\item{header}{logical. Does the data set contain column names?}

\item{...}{further arguments to be passed to \code{bigmemory::read.big.matrix}.}
}
\description{
read genotype data from file to bigmemory object, which can be passed to a  GWASdata object
}
\details{
If the data set contains rownames specified, set option \code{has.row.names = TRUE}.
}
\seealso{
\code{\link{GWASdata-class}}
}

