% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rs10243170_info}
\alias{rs10243170_info}
\title{Example \code{snp_info} object for SNP rs10243170.}
\format{A \code{snp_info} object including information on the SNP as
extracted from the Ensembl database.
\describe{
 \item{info}{a \code{data frame} including the extracted information on the
SNP. Columns given are 'chr', 'position', and 'rsnumber'}
}}
\source{
Ensembl extract
}
\usage{
rs10243170_info
}
\description{
An object of class \code{snp_info} for rs10243170.
}
\keyword{datasets}

