% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{sample_genes,pathway-method}
\alias{sample_genes,pathway-method}
\title{function randomly selecting effect genes in pathway}
\usage{
\S4method{sample_genes}{pathway}(object, no = 3)
}
\arguments{
\item{no}{a numeric constant specifying the number of genes to be sampled, default is 3}

\item{pathway}{pathway object}
}
\value{
a vector of length \code{no} with vertex id's of sampled genes
}
\description{
function randomly selecting effect genes in pathway
}
\examples{
data(hsa04710)
summary(hsa04710)
sample3 <- sample_genes(hsa04710, no = 3)
plot(hsa04710, highlight.genes = sample3)
sample10 <- sample_genes(hsa04710, no = 10)
plot(hsa04710, highlight.genes = sample10)
}
\author{
Juliane Manitz
}

