% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lkmt.r
\docType{methods}
\name{score_test,matrix-method}
\alias{score_test,matrix-method}
\title{Calculates the p-value for a kernelmatrix using Satterthwaite approximation}
\usage{
\S4method{score_test}{matrix}(x1, x2)
}
\arguments{
\item{x1}{A \code{\link{matrix}} which is the
similarity matrix calculated for the pathway to be tested.}

\item{x2}{An \code{lm} or \code{glm} object of the nullmodel with fixed
effects covariates included, but no genetic random effects.}
}
\value{
A \code{list} including the test results of the pathway.
}
\description{
This function evaluates a pathways influence on an individuals probability
of beeing a case using the logistic kernel machine test. P-values are
determined using a Sattherthwaite Approximation as described by Dan Schaid.
}
\author{
Stefanie Friedrichs, Saskia Freytag, Ngoc-Thuy Ha
\itemize{
\item Schaid DJ: Genomic Similarity and Kernel Methods I: Advancements by
Building on Mathematical and Statistical Foundations. Hum Hered 2010, 70:109-31
}
}

