% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lkmt.r
\name{score_test}
\alias{score_test}
\title{Calculates the p-value for a kernelmatrix}
\usage{
score_test(kernels, nullmodel, pd.check = TRUE)
}
\arguments{
\item{kernels}{an object of \code{\link{kernel}} class. Includes the
similarity matrix calculated for the pathway to be tested.}

\item{nullmodel}{A \code{glm} object of the nullmodel with fixed effects
covariates included, but no genetic random effects.}

\item{pd.check}{boolean, whether to check for positive definiteness.}
}
\value{
A \code{list} including the test results of the pathway.
}
\description{
This function evaluates a pathways influence on an individuals probability
of beeing a case using the logistic kernel machine test. P-values are
determined using a Sattherthwaite Approximation as described by Dan Schaid.
}
\author{
Stefanie Friedrichs, Saskia Freytag, Ngoc-Thuy Ha
For details on the p-value approximation see
\itemize{
\item Schaid DJ: Genomic Similarity and Kernel Methods I: Advancements by Building on Mathematical and Statistical Foundations. Hum Hered 2010, 70:109-31
}
}

