% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\docType{methods}
\name{set_one,matrix-method}
\alias{set_one,matrix-method}
\title{Helper function to set matrix entries to 0/1/-1 only}
\usage{
\S4method{set_one}{matrix}(x)
}
\arguments{
\item{x}{numeric \code{matrix} representing the network adjacency matrix.}
}
\value{
A \code{matrix} representing the interaction network in the pathway
with entries equal to 1, -1 or 0.
}
\description{
This function sets all entries in a matrix bigger than 1 to 1 and all entries
smaller than -1 to -1. It is called by \code{get_network_matrix}.
(For internal use)
}
\author{
Stefanie Friedrichs
}
\keyword{internal}

