% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\name{set_one}
\alias{set_one}
\title{Helper function to set matrix entries to 0/1/-1 only}
\usage{
set_one(M)
}
\arguments{
\item{M}{A \code{matrix} representing the networkmatrix.}
}
\value{
A \code{matrix} representing the interaction network in the pathway
with entries equal to 1, -1 or 0.
}
\description{
This function sets all entries in a matrix bigger than 1 to 1 and all entries
smaller than -1 to -1. It is called by \code{\link{get_network_matrix}}.
}
\author{
Stefanie Friedrichs
}

