% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{show,GWASdata-method}
\alias{show,GWASdata-method}
\title{Shows basic information on \code{GWASdata} object}
\usage{
\S4method{show}{GWASdata}(object)
}
\arguments{
\item{gwas}{\code{GWASdata} object.}
}
\value{
This function shows the phenotype information and the object description included in a \code{\link{GWASdata}} object.
}
\description{
Shows basic information on \code{GWASdata} object
}
\examples{
data(gwas)
show(gwas)
}
\author{
Juliane Manitz
}

