% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lkmt.r
\docType{methods}
\name{show,lkmt-method}
\alias{show,lkmt-method}
\title{Shows basic information on \code{lkmt} object}
\usage{
\S4method{show}{lkmt}(object)
}
\arguments{
\item{lkmt}{An object of class \code{\link{lkmt}}.}
}
\description{
Shows basic information on \code{lkmt} object
}
\author{
Juliane Manitz
}

