% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\name{snp_info}
\alias{snp_info}
\title{Get SNP positions}
\usage{
snp_info(snps, ...)
}
\arguments{
\item{snps}{A vector of SNP rsnumbers for which positions will be extracted.}
}
\value{
A \code{data.frame} including the SNP positions with columns
chromosome, position and rsnumber. SNPs not found in the Ensemble database
not be listed in the returned \code{data.frame}, SNPs with multiple positions
will appear several times.
}
\description{
This function gives for a vector of SNPs the position of each SNP as
extracted from the Ensemble database. The database is accessed via the
R-package \code{biomaRt}.
}
\examples{
# snp_info("rs234") ### ERROR
}
\author{
Stefanie Friedrichs
}

