% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\name{pathway_info}
\alias{pathway_info}
\title{An S4 class for an object assigning genes to pathways}
\usage{
pathway_info(x, ...)
}
\description{
An S4 class for an object assigning genes to pathways
}
\section{Slots}{

\describe{
\item{\code{info}}{A \code{data.frame} including information on genes contained in
pathways}
}}
\examples{
data(hsa04022_info)
}
\author{
Stefanie Friedrichs
}

