% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{summary,GWASdata-method}
\alias{summary,GWASdata-method}
\title{Summarizes the content of a \code{GWASdata} object}
\usage{
\S4method{summary}{GWASdata}(object)
}
\arguments{
\item{gwas}{\code{GWASdata} object.}
}
\value{
This function gives an overview about the information included in a \code{\link{GWASdata}} object. Summary statistics for phenotype and genotype data are calculated.
}
\description{
Summarizes the content of a \code{GWASdata} object
}
\examples{
# data(gwas)   #### ERROR
# summary(gwas) ### ERROR
}
\author{
Juliane Manitz
}

