% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lkmt.r
\docType{methods}
\name{summary,lkmt-method}
\alias{summary,lkmt-method}
\title{Summarizes information on \code{lkmt} object}
\usage{
\S4method{summary}{lkmt}(object)
}
\arguments{
\item{lkmt}{An object of class \code{\link{lkmt}}.}
}
\description{
Summarizes information on \code{lkmt} object
}
\author{
Juliane Manitz
}

