% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\docType{methods}
\name{calc_kernel,GWASdata-method}
\alias{calc_kernel,GWASdata-method}
\alias{lin_kernel,GWASdata-method}
\alias{net_kernel,GWASdata-method}
\alias{sia_kernel,GWASdata-method}
\title{calculates a kernel to be evaluated in the logistic kernel machine test.}
\usage{
\S4method{calc_kernel}{GWASdata}(object, pathway, knots = NULL,
  type = c("lin", "sia", "net"), parallel = c("none", "cpu", "gpu"), ...)

\S4method{lin_kernel}{GWASdata}(object, pathway, knots = NULL,
  parallel = c("none", "cpu", "gpu"), ...)

\S4method{sia_kernel}{GWASdata}(object, pathway, knots = NULL,
  parallel = c("none", "cpu", "gpu"), ...)

\S4method{net_kernel}{GWASdata}(object, pathway, knots = NULL,
  parallel = c("none", "cpu", "gpu"), ...)
}
\arguments{
\item{object}{\code{GWASdata} object containing the genotypes of the individuals for which a kernel will be calculated.}

\item{pathway}{object of the class \code{pathway} specifying the SNP set for which a kernel will be calculated.}

\item{knots}{\code{GWASdata} object, if specified low-rank kernel be computed}

\item{type}{character, kernel type: Use \code{"lin"} for linear kernels, \code{"sia"} for size-adjusted or \code{"net"} for network-based kernels.}

\item{parallel}{character specifying if the kernel matrix is computed in parallel: Use \code{"none"} for ... <FIXME>}

\item{...}{further arguments to be passed to the kernel computations}
}
\value{
Returns object of class kernel, including the similarity matrix of the corresponding pathway for the considered individuals (\code{kernel}). If \code{knots} are specified low-rank kernel of class \code{lowrank_kernel} will be returned, which is not necessarily quadratic and symmetric.
}
\description{
calculates a kernel to be evaluated in the logistic kernel machine test.
}
\details{
Different types of kernels can be constructed:
\itemize{
  \item \code{type='lin'} creates the linear kernel assuming additive SNP effects to be evaluated in the logistic kernel machine test.
  \item \code{type='sia'} calculates the size-adjusted kernel which takes into consideration the numbers of SNPs and genes in a pathway to correct for size bias.
  \item \code{type='net'} calculates a network-based kernel to be evaluated in the logistic kernel machine test. Not only information on gene membership and gene/pathway size in number of SNPs is incorporated in the calculation, but also the interaction structure of genes in the pathway.
}
For more details, check the references.
}
\section{Methods (by class)}{
\itemize{
\item \code{GWASdata}: 

\item \code{GWASdata}: 

\item \code{GWASdata}: 
}}
\examples{
#### missing example #### <FIXME>
}
\author{
Juliane Manitz, Saskia Freytag, Ngoc Thuy Ha
}
\references{
\itemize{
 \item Wu MC, Kraft P, Epstein MP, Taylor DM, Chanock SJ, Hunter DJ, Lin X Powerful SNP-Set Analysis for Case-Control Genome-Wide Association Studies. Am J Hum Genet 2010, 86:929-42
 \item Freytag S, Bickeboeller H, Amos CI, Kneib T, Schlather M: A Novel Kernel for Correcting Size Bias in the Logistic Kernel Machine Test with an Application to Rheumatoid Arthritis. Hum Hered. 2012, 74(2):97-108.
 \item Freytag S, Manitz J, Schlather M, Kneib T, Amos CI, Risch A, Chang-Claude J, Heinrich J, Bickeboeller H: A network-based kernel machine test for the identification of risk pathways in genome-wide association studies. Hum Hered. 2013, 76(2):64-75.
}
}
\seealso{
\code{\link{kernel-class}}, \code{\link{GWASdata-class}}, \code{\link{pathway-class}}
}

