\name{GeneSNPsize-methods}
\docType{methods}
\alias{GeneSNPsize-methods}
\alias{GeneSNPsize,GWASdata-method}
\title{Methods for Function \code{GeneSNPsize}}
\description{
Methods for function \code{GeneSNPsize}
}
\section{Methods}{
\describe{
Function that counts the number of genes and SNPs in each pathway. 
}
\usage{
GeneSNPsize(object) {
}
\arguments{
  \item{object}{
  object of class GWASdata, including the annotation file on which gene and SNP numbers will be counted.
}
}
}
\keyword{methods}
