% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gwas}
\alias{gwas}
\title{Example \code{\link{GWASdata}} object.}
\format{An object of class \code{\link{GWASdata}}:
\describe{
 \item{geno}{contains example genotypes}
 \item{anno}{example annotation for three pathways}
 \item{pheno}{exemplary phenotypes for all 'genotyped' individuals}
 \item{desc}{a description of the GWAS study, here 'example study'}
}}
\source{
simulated data
}
\usage{
data(gwas)
}
\description{
An object of type GWASdata containing the example files for 
annotation, phenotypes and genotypes.
}
\examples{
# create gwas object
data(pheno)
data(geno)
data(anno)
gwas <- new('GWASdata', pheno=pheno, geno=geno, anno=anno, desc="some study") 
}
\keyword{datasets}
