% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{read_geno,ANY-method}
\alias{GeneSNPsize,GWASdata,ANY-method}
\alias{GeneSNPsize,GWASdata-method}
\alias{read_geno,ANY-method}
\alias{show,GWASdata,ANY-method}
\alias{show,GWASdata-method}
\alias{summary,GWASdata,ANY-method}
\alias{summary,GWASdata-method}
\title{read genotype data from file to ff_matrix object, which can be passed to a  GWASdata object}
\usage{
\S4method{read_geno}{ANY}(path)

\S4method{show}{GWASdata}(object)

\S4method{summary}{GWASdata}(object)

\S4method{GeneSNPsize}{GWASdata}(object)
}
\arguments{
\item{path}{character, which contains the path to the txt file}

\item{object}{A \code{GWASdata} object.}

\item{object}{A \code{GWASdata} object.}

\item{object}{A \code{GWASdata} object.}
}
\value{
\code{show} This function shows the phenotype information and the object description included in a \code{\link{GWASdata}} object.

\code{summary} This function gives an overview about the information included in a \code{\link{GWASdata}} object. Summary statistics for phenotype and genotype data are calculated.

\code{GeneSNPsize} Creates a list of pathway names with numbers of snps and genes in the pathway.
}
\description{
read genotype data from file to ff_matrix object, which can be passed to a  GWASdata object

\code{show} Shows basic information on \code{GWASdata} object

\code{summary} Summarizes the content of a \code{GWASdata} object

\code{GeneSNPsize} Counts the number of Genes and SPNs in each pathway
}
\examples{
data(gwas)
show(gwas)

#
# data(gwas)   #### ERROR
# summary(gwas) ### ERROR

#
#data(gwas)### ERROR
#GeneSNPsize(gwas) ### ERROR

#
}
\author{
Juliane Manitz

Juliane Manitz

Juliane Manitz
}

