% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{class}
\name{GWASdata-class}
\alias{GWASdata}
\alias{GWASdata-class}
\title{An S4 class defining an object to represent a Genome-wide Assocaition Study.}
\description{
An S4 class defining an object to represent a Genome-wide Assocaition Study.
}
\section{Slots}{

\describe{
\item{\code{pheno}}{A \code{data.frame} specifying individual IDs, phenotypes and   covariates to be included in the regression model e.g. ID, pheno, sex, pack.years. Note: IDs have to be in the first column!}
}}
\examples{
data(pheno)
data(geno)
#gwas <- new('GWASdata', pheno=pheno, geno=geno, desc="some study") ### ERROR
}
\author{
Juliane Manitz, Stefanie Friedrichs
}
\seealso{
\code{\link{}}
}

