% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{GWASdata}
\alias{GWASdata}
\alias{GWASdata,ANY-method}
\alias{GeneSNPsize,GWASdata-method}
\alias{show,GWASdata-method}
\alias{summary,GWASdata-method}
\title{An S4 class defining an object to represent a Genome-wide Assocaition Study.}
\usage{
GWASdata(object, ...)

\S4method{GWASdata}{ANY}(geno, anno, pheno = NULL, desc = "")

\S4method{show}{GWASdata}(object)

\S4method{summary}{GWASdata}(object)

\S4method{GeneSNPsize}{GWASdata}(object)
}
\arguments{
\item{object}{A \code{GWASdata} object.}
}
\description{
An S4 class defining an object to represent a Genome-wide Assocaition Study.

\code{'GWASdata'} is a GWASdata object constructor

\code{show} displays basic information on \code{GWASdata} object

\code{summary} summarizes the content of a \code{GWASdata} object and gives an overview about the information included in a \code{\link{GWASdata}} object. Summary statistics for phenotype and genotype data are calculated.

\code{GeneSNPsize} creates a \code{data.frame} of pathway names with numbers of snps and genes in each pathway.
}
\section{Slots}{

\describe{
\item{\code{geno}}{an object of any type, including genotype information.}

\item{\code{anno}}{A \code{data.frame} mapping SNPs to genes and genes to
pathways. Needs to include the columns 'pathway' (pathway ID, e.g. hsa
number from KEGG database), 'gene' (gene name (hgnc_symbol)), 'chr'
(chromosome), 'snp' (rsnumber) and 'position' (base pair position of SNP).}

\item{\code{pheno}}{A \code{data.frame} specifying individual IDs, phenotypes and
covariates to be included in the regression model e.g. ID, pheno, sex,
pack.years. Note: IDs have to be in the first column!}

\item{\code{desc}}{A \code{character} giving the GWAS description, e.g. name of study.  Default is ''.}
}}
\examples{
data(pheno)
data(geno)
data(anno)
gwas <- new('GWASdata', pheno=pheno, geno=geno, anno=anno, desc="some study")
data(gwas)
gwas
data(gwas)
summary(gwas)
data(gwas)
GeneSNPsize(gwas)
}
\author{
Juliane Manitz, Stefanie Friedrichs

Juliane Manitz
}

