% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{class}
\name{GWASdata-class}
\alias{GWASdata}
\alias{GWASdata-class}
\alias{GeneSNPsize,GWASdata,ANY-method}
\alias{GeneSNPsize,GWASdata-method}
\alias{show,GWASdata,ANY-method}
\alias{show,GWASdata-method}
\alias{summary,GWASdata,ANY-method}
\alias{summary,GWASdata-method}
\title{An S4 class defining an object to represent a Genome-wide Assocaition Study.}
\usage{
\S4method{show}{GWASdata}(object)

\S4method{summary}{GWASdata}(object)

\S4method{GeneSNPsize}{GWASdata}(object)
}
\arguments{
\item{object}{\code{GWASdata} object.}

\item{object}{\code{GWASdata} object.}

\item{gwas}{\code{GWASdata} object.}
}
\value{
This function shows the phenotype information and the object
description included in a \code{\link{GWASdata}} object.

This function gives an overview about the information included in a
\code{\link{GWASdata}} object. Summary statistics for phenotype and genotype
data are calculated.

Creates a list of pathway names with numbers of snps and genes in the pathway.
}
\description{
An S4 class defining an object to represent a Genome-wide Assocaition Study.

Shows basic information on \code{GWASdata} object

Summarizes the content of a \code{GWASdata} object

Counts the number of Genes and SPNs in each pathway
}
\section{Slots}{

\describe{
\item{\code{pheno}}{A \code{data.frame} specifying individual IDs, phenotypes and
covariates to be included in the regression model e.g. ID, pheno, sex,
pack.years. Note: IDs have to be in the first column!}

\item{\code{geno}}{An \code{ffdf} data frame including genotype information.}

\item{\code{anno}}{A \code{data.frame} with mapping SNPs to genes and genes to pathways.
Needs to include the columns 'pathway' (pathway ID, e.g. hsa number from KEGG
database), 'gene' (gene name (hgnc_symbol)), 'chr' (chromosome), 'snp'
(rsnumber) and 'position' (base pair position of SNP).}

\item{\code{desc}}{A \code{character} giving the GWAS description, e.g. name of study}
}}
\examples{
data(pheno)
data(geno)
#gwas <- new('GWASdata', pheno=pheno, geno=geno, desc="some study") ### ERROR
data(gwas)
show(gwas)
# data(gwas)   #### ERROR
# summary(gwas) ### ERROR
#data(gwas)### ERROR
#GeneSNPsize(gwas) ### ERROR
}
\author{
Juliane Manitz, Stefanie Friedrichs

Juliane Manitz

Juliane Manitz

Juliane Manitz
}
\seealso{
\code{\link{}}
}

