% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{GeneSNPsize}
\alias{GeneSNPsize}
\alias{GeneSNPsize,GWASdata,ANY-method}
\title{\code{GeneSNPsize} creates a \code{data.frame} of pathway names with numbers of snps and genes in each pathway.}
\usage{
\S4method{GeneSNPsize}{GWASdata}(object)
}
\description{
\code{GeneSNPsize} creates a \code{data.frame} of pathway names with numbers of snps and genes in each pathway.
}
\examples{
# SNPs and genes in pathway
data(gwas)
GeneSNPsize(gwas)
}

