% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GWASdata.r
\docType{methods}
\name{GWASdata}
\alias{GWASdata}
\alias{GWASdata,ANY-method}
\alias{show,GWASdata,ANY-method}
\title{An S4 class defining an object to represent a Genome-wide Assocaition Study.}
\usage{
GWASdata(object, ...)

\S4method{GWASdata}{ANY}(geno, anno, pheno = NULL, desc = "")
}
\arguments{
\item{geno}{an \code{ff} data frame including genotype information.}

\item{anno}{a \code{data.frame} mapping SNPs to genes and genes to
pathways. Needs to include the columns 'pathway' (pathway ID, e.g. hsa
number from KEGG database), 'gene' (gene name (hgnc_symbol)), 'chr'
(chromosome), 'snp' (rsnumber) and 'position' (base pair position of SNP).}

\item{pheno}{A \code{data.frame} specifying individual IDs, phenotypes and
covariates to be included in the regression model e.g. ID, pheno, sex,
pack.years. Note: IDs have to be in the first column. Default is \code{NULL}.}

\item{desc}{A \code{character} giving the GWAS description, e.g. name of study. Defaul is ''.

#}
}
\description{
An S4 class defining an object to represent a Genome-wide Assocaition Study.

\code{'GWASdata'} is a GWASdata object constructor
}
\section{Slots}{

\describe{
\item{\code{geno}}{An \code{ffdf} data frame including genotype information.}

\item{\code{anno}}{A \code{data.frame} mapping SNPs to genes and genes to
pathways. Needs to include the columns 'pathway' (pathway ID, e.g. hsa
number from KEGG database), 'gene' (gene name (hgnc_symbol)), 'chr'
(chromosome), 'snp' (rsnumber) and 'position' (base pair position of SNP).}

\item{\code{pheno}}{A \code{data.frame} specifying individual IDs, phenotypes and
covariates to be included in the regression model e.g. ID, pheno, sex,
pack.years. Note: IDs have to be in the first column!}

\item{\code{desc}}{A \code{character} giving the GWAS description, e.g. name of study}
}}
\examples{
data(pheno)
data(geno)
#gwas <- new('GWASdata', pheno=pheno, geno=geno, desc="some study") ### ERROR
}
\author{
Juliane Manitz, Stefanie Friedrichs

Juliane Manitz
}

