% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsa04020}
\alias{hsa04020}
\title{Example \code{pathway} object for pathway hsa04020.}
\format{A \code{pathway} object including 180 genes.
\describe{
 \item{id}{KEGG identifier of the example pathways}
 \item{adj}{gives the quadratic adjacency matrix for the pathway and with 
that the network topology. Matrix dimensionequal the number of genes in the 
pathway}
 \item{sign}{includes a vector of signs to distinguish activations and 
inhibitions in the adjacency matrix }
}}
\source{
simulated data and Ensembl extract
}
\usage{
data(hsa04020)
}
\description{
An object of class \code{pathway} for the pathway with KEGG identifier hsa04020.
}
\keyword{datasets}
