% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsa04022_info}
\alias{hsa04022_info}
\title{Example \code{pathway_info} object for pathway hsa04022.}
\format{A \code{pathway_info} object including information on 163 genes.
\describe{
 \item{info}{a \code{data frame} including information on genes included in
pathway. Has columns 'pathway', 'gene_start', 'gene_end', 'chr', and 'gene'}
}}
\source{
Ensembl extract
}
\usage{
hsa04022_info
}
\description{
An object of class \code{pathway_info} for the pathway with KEGG
identifier hsa04020.
}
\keyword{datasets}

