% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\docType{methods}
\name{kernel}
\alias{kernel}
\alias{kernel,ANY-method}
\alias{lin_kernel,ANY-method}
\alias{net_kernel,ANY-method}
\alias{plot,kernel,ANY-method}
\alias{plot,kernel,missing-method}
\alias{show,kernel,ANY-method}
\alias{show,kernel-method}
\alias{sia_kernel,ANY-method}
\alias{summary,kernel,ANY-method}
\alias{summary,kernel-method}
\title{An S4 class to represent a kernel for a SNPset}
\usage{
kernel(object, ...)

\S4method{show}{kernel}(object)

\S4method{summary}{kernel}(object)

\S4method{plot}{kernel,missing}(x, y = NA, hclust = FALSE, ...)

\S4method{kernel}{ANY}(type = c("lin", "sia", "net"), data, pathway,
  parallel = c("none", "cpu", "gpu"), ...)

\S4method{lin_kernel}{ANY}(data, pathway, parallel = c("none", "cpu", "gpu"),
  ...)

\S4method{sia_kernel}{ANY}(data, pathway, parallel = c("none", "cpu", "gpu"),
  ...)

\S4method{net_kernel}{ANY}(data, pathway, parallel = c("none", "cpu", "gpu"),
  ...)
}
\arguments{
\item{object}{kernel object}

\item{y}{missing (placeholder)}

\item{hclust}{\code{logical}, indicating whether a dendrogram should be added}

\item{type}{character, kernel type: Use \code{"lin"} for linear kernels, \code{"sia"} for size-adjusted or \code{"net"} for network-based kernels.}

\item{data}{Object of the class \code{GWASdata} containing the genotypes of the individuals for which a kernel will be calculated.}

\item{pathway}{Object of the class \code{pathway} specifying the SNP set for which a kernel will be calculated.}
}
\value{
Returns object of class kernel, including the similarity matrix of the corresponding pathway for the considered individuals (\code{kernel}).
}
\description{
An S4 class to represent a kernel for a SNPset

\code{show} displays the kernel object briefly

\code{summary} generates a kernel object summary including the number of individuals and genes for the pathway

\code{plot} creates an image plot of a kernel object

\code{kernel} is a kernel object constructor and creates a kernel to be evaluated in the logistic kernel machine test.
}
\details{
Different types of kernels can be constructed:
\itemize{
  \item \code{type='lin'} creates the linear kernel assuming additive SNP effects to be evaluated in the logistic kernel machine test.
  \item \code{type='sia'} calculates the size-adjusted kernel which takes into consideration the numbers of SNPs and genes in a pathway to correct for size bias.
  \item \code{type='net'} calculates a network-based kernel to be evaluated in the logistic kernel machine test. Not only information on gene membership and gene/pathway size in number of SNPs is incorporated in the calculation, but also the interaction structure of genes in the pathway.
}
For more details, check the references.
}
\section{Methods (by generic)}{
\itemize{
\item \code{lin_kernel}: 

\item \code{sia_kernel}: 

\item \code{net_kernel}: 
}}
\section{Slots}{

\describe{
\item{\code{type}}{character, kernel type: Use \code{"lin"} for linear kernels,
\code{"sia"} for size-adjusted or \code{"net"} for network-based kernels.}

\item{\code{kernel}}{kernel matrix of dimension equal to individuals}

\item{\code{GWASdata}}{GWASdata object}

\item{\code{pathway}}{pathway object}
}}
\author{
Juliane Manitz

Juliane Manitz, Saskia Freytag, Ngoc Thuy Ha
}
\references{
\itemize{
 \item Wu MC, Kraft P, Epstein MP, Taylor DM, Chanock SJ, Hunter DJ, Lin X Powerful SNP-Set Analysis for Case-Control Genome-Wide Association Studies. Am J Hum Genet 2010, 86:929-42
 \item Freytag S, Bickeboeller H, Amos CI, Kneib T, Schlather M: A Novel Kernel for Correcting Size Bias in the Logistic Kernel Machine Test with an Application to Rheumatoid Arthritis. Hum Hered. 2012, 74(2):97-108.
 \item Freytag S, Manitz J, Schlather M, Kneib T, Amos CI, Risch A, Chang-Claude J, Heinrich J, Bickeboeller H: A network-based kernel machine test for the identification of risk pathways in genome-wide association studies. Hum Hered. 2013, 76(2):64-75.
}
}
\seealso{
\code{\link{GWASdata-class}}, \code{\link{pathway-class}}
}

