% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\docType{class}
\name{kernel-class}
\alias{kernel}
\alias{kernel-class}
\alias{plot,kernel,ANY-method}
\alias{plot,kernel,missing-method}
\alias{show,kernel-method}
\alias{summary,kernel,ANY-method}
\alias{summary,kernel-method}
\title{An S4 class representing the kernel of a pathway}
\usage{
\S4method{show}{kernel}(object)

\S4method{summary}{kernel}(object)

\S4method{plot}{kernel,missing}(x, y = NA, hclust = FALSE, ...)
}
\arguments{
\item{object}{kernel object}

\item{y}{missing (placeholder)}

\item{hclust}{\code{logical}, indicating whether a dendrogram should be added}
}
\description{
An S4 class representing the kernel of a pathway

\code{show} displays the kernel object briefly

\code{summary} generates a kernel object summary including the number of individuals and genes for the pathway

\code{plot} creates an image plot of a kernel object
}
\section{Slots}{

\describe{
\item{\code{type}}{A \code{character} representing the kernel type: Use
\code{"lin"} for linear kernels, \code{"sia"} for size-adjusted
or \code{"net"} for network-based kernels.}

\item{\code{kernel}}{A kernel matrix of dimension equal to the number of individuals}

\item{\code{pathway}}{A pathway object}
}}
\author{
Juliane Manitz
}

