% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lkmt.r
\docType{methods}
\name{lkmt}
\alias{lkmt}
\alias{show,GWASdata,ANY-method}
\alias{show,lkmt-method}
\alias{summary,GWASdata,ANY-method}
\alias{summary,lkmt-method}
\title{An S4 class to represent the variance component test.}
\usage{
lkmt(formula, kernel, GWASdata, method = c("satt", "davies"), ...)

\S4method{show}{lkmt}(object)

\S4method{summary}{lkmt}(object)
}
\arguments{
\item{object}{An object of class \code{\link{lkmt}}.}

\item{object}{An object of class \code{\link{lkmt}}.}
}
\value{
\code{show} Basic information on \code{lkmt} object.
#

\code{summary} Summarized information on \code{lkmt} object.
#
}
\description{
An S4 class to represent the variance component test.

\code{show} Shows basic information on \code{lkmt} object

\code{summary} Summarizes information on \code{lkmt} object
}
\section{Slots}{

\describe{
\item{\code{formula}}{A formular stating the regression nullmodel that will be used in
the variance component test.}

\item{\code{kernel}}{An object of class \code{\link{kernel}} representing the similarity
matrix of the individuals based on which the pathways influence is evaluated.}

\item{\code{GWASdata}}{An object of class \code{\link{GWASdata}} stating the data on
which the test is conducted.}

\item{\code{statistic}}{A \code{vector} giving the value of the variance component
ttest statistic.}

\item{\code{statistic}}{A \code{vector} giving the number of degrees of freedom.}

\item{\code{statistic}}{A \code{vector} giving the p-value calculated for the pathway
in the variance component test.}
}}
\author{
Juliane Manitz, Stefanie Friedrichs

Juliane Manitz

Juliane Manitz
}

