% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lkmt.r
\name{lkmt}
\alias{lkmt}
\title{Function to start the logistic kernel machine test.}
\usage{
lkmt(formula, kernel, GWASdata, method = c("satt", "davies"), ...)
}
\arguments{
\item{formula}{A formular for the regression nullmodel that will be used in
the variance component test.}

\item{kernel}{An object of class \code{\link{kernel}} representing the similarity
matrix of the individuals based on which the pathways influence is evaluated.}

\item{GWASdata}{An object of class \code{\link{GWASdata}} representing the data on
which the test is conducted.}

\item{method}{A \code{character} specifying how p-values should be calculated.
For Satterthwaite approximation use 'satt', for Davies method use 'davies'.}

\item{...}{additional arguments can be added.}
}
\value{
An \code{\link{lkmt}} object with the test results.
}
\description{
Function to start the logistic kernel machine test.
}
\author{
Juliane Manitz, Stefanie Friedrichs

For details on the variance component test see the references.
}
\references{
\itemize{
 \item Wu MC, Kraft P, Epstein MP, Taylor DM, Chanock SJ, Hunter DJ, Lin X: Powerful SNP-Set Analysis for Case-Control Genome-Wide Association Studies. Am J Hum Genet 2010, 86:929-42
}
}

