% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lkmt.r
\name{lkmt}
\alias{lkmt}
\title{An S4 class defining an object to represent the variance component test.}
\usage{
lkmt(formula, kernel, GWASdata, ...)

lkmt(formula, kernel, GWASdata, ...)
}
\value{
An \code{\link{lkmt}} object with the test results.
}
\description{
the variance component test.
matrix of the individuals based on which the pathways influence is evaluated.
which the test is conducted.
ttest statistic.
in the variance component test.

the variance component test.
matrix of the individuals based on which the pathways influence is evaluated.
which the test is conducted.
}
\author{
Juliane Manitz, Stefanie Friedrichs

Juliane Manitz, Stefanie Friedrichs

For details on the variance component test see
}
\references{
\itemize{
 \item Wu MC, Kraft P, Epstein MP, Taylor DM, Chanock SJ, Hunter DJ, Lin X: Powerful SNP-Set Analysis for Case-Control Genome-Wide Association Studies. Am J Hum Genet 2010, 86:929-42
}
}

