% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\docType{class}
\name{lowrank_kernel-class}
\alias{lowrank_kernel}
\alias{lowrank_kernel-class}
\title{An S4 class to represent a low-rank kernel for a SNPset at specified knots}
\description{
An S4 class to represent a low-rank kernel for a SNPset at specified knots
}
\details{
This kernel is used for predictions. If observations and knots are equal, better construct a full-rank kernel of class \code{\link{kernel}}.
}
\section{Slots}{

\describe{
\item{\code{type}}{character, kernel type: Use \code{"lin"} for linear kernels,
\code{"sia"} for size-adjusted or \code{"net"} for network-based kernels.}

\item{\code{kernel}}{kernel matrix of dimension equal to individuals}

\item{\code{pathway}}{pathway object}
}}
\author{
Juliane Manitz
}

