% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{net.kernel.hsa04020}
\alias{net.kernel.hsa04020}
\title{Example network-based kernel matrix for pathway hsa04020.}
\format{An object of class \code{kernel} and type 'network' for the pathway
hsa04020.
\describe{
 \item{type}{specifies which kernelfunction was used to calculate the kernel}
\item{kernel}{includes the kernelmatrix calculated for the pathway}
\item{pathway}{includes the pathway object of the pathway, for which the
kernelmatrix was calculated}
}}
\source{
simulated data and Ensembl extract
}
\usage{
data(net.kernel.hsa04020)
}
\description{
An example of a kernel object.
}
\keyword{datasets}

