% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\docType{methods}
\name{rewire_network,matrix-method}
\alias{rewire_network,matrix-method}
\title{Apply two-step network to rewire network genes if it contains no SNPs in GWASdata (for internal use)}
\usage{
\S4method{rewire_network}{matrix}(x, remov)
}
\arguments{
\item{x}{adjacency matrix}

\item{remov}{indication which genes should be removed <FIXME> what does this mean? TRUE/FALSE vector, gene names, or vector}
}
\value{
An adjacency matrix containing rewired network
}
\description{
Apply two-step network to rewire network genes if it contains no SNPs in GWASdata (for internal use)
}
\author{
Juliane Manitz
}
\references{
TODO Newman?
}

