% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\docType{methods}
\name{rewire_network,matrix-method}
\alias{rewire_network,matrix-method}
\title{Rewires interactions in a pathway, which go through a gene not represented
by any SNPs in the considered \code{GWASdata} object. (for internal use)}
\usage{
\S4method{rewire_network}{matrix}(x, remov)
}
\arguments{
\item{x}{Adjacency \code{matrix}}

\item{remov}{A \code{vector} of gene names, indicating which genes are not
represented by SNPs in the considered \code{GWASdata} and will be removed}
}
\value{
An adjacency \code{matrix} containing the rewired network
}
\description{
Rewires interactions in a pathway, which go through a gene not represented
by any SNPs in the considered \code{GWASdata} object. (for internal use)
}
\author{
Juliane Manitz
}

