% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kernel.r
\name{rewire_network}
\alias{rewire_network}
\title{Apply two-step network to rewire network genes if it contains no SNPs in GWASdata (for internal use)}
\usage{
rewire_network(N, remov)
}
\arguments{
\item{N}{adjacency matrix}

\item{remov}{indication which genes should be removed}
}
\value{
An adjacency matrix containing rewired network
}
\description{
Apply two-step network to rewire network genes if it contains no SNPs in GWASdata (for internal use)
}
\author{
Juliane Manitz
}
\references{
TODO Newman?
}

