% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kernel.r
\docType{methods}
\name{rewire_network}
\alias{rewire_network}
\alias{rewire_network,pathway-method}
\title{Rewires interactions in a pathway, which go through a gene not represented 
by any SNPs in the considered GWAS dataset.}
\usage{
\S4method{rewire_network}{pathway}(object, x)
}
\arguments{
\item{object}{\code{pathway} object which's network matrix will be rewired}

\item{x}{A \code{vector} of gene names, indicating which genes are not 
represented by SNPs in the considered \code{GWASdata} and will be removed}
}
\value{
A \code{pathway} object including the rewired network
}
\description{
Rewires interactions in a pathway, which go through a gene not represented 
by any SNPs in the considered GWAS dataset.
}
\examples{
\dontrun{ 
data(hsa04020)
rewire_network(hsa04020, c("PHKB", "ORAI2"))
}
}
\author{
Stefanie Friedrichs, Juliane Manitz
}
