% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pathway.r
\name{set_names}
\alias{set_names}
\title{Helper function to translate gene numbers to names}
\usage{
set_names(N, nodes, my_list)
}
\arguments{
\item{N}{A \code{matrix} representing the networkmatrix.}

\item{nodes}{A \code{vector} of gene numbers to be replaced by names.}

\item{my_list}{A \code{data.frame} listing gene names and numbers. Output from
\code{gene_name_number}.}
}
\value{
A \code{matrix} representing the interaction network in the pathway
with gene names as rownames and columnnames.
}
\description{
This function exchanges the numbers used for genes in KEGG download KGML files
with the corresponding gene names. Names are set to be the columnnames and
rownames of the pathway's networkmatrix. The function
is called by \code{\link{get_network_matrix}}. (For internal use)
}
\author{
Stefanie Friedrichs
}

